--- semana19/aula1/src/data/user/updateUser.ts
@@ -1,18 +1,0 @@
-import { connection } from "../connection";
-
-export default async function updateUser(
-   id: string,
-   name?: string,
-   nickname?: string,
-   email?: string
-) {
-
-   await connection("to_do_list_users")
-      .update({
-         name, nickname, email
-      })
-      .where({ 
-         id 
-      })
-
-}
