+++ semana19/aula1/src/index.ts
@@ -1,16 +1,22 @@
 import express from 'express'
+import knex from 'knex'
 import cors from 'cors'
 import dotenv from 'dotenv'
-import createUser from './controller/user/createUser'
-import getUserById from './controller/user/getUserById'
-import editUser from './controller/user/editUser'
-import createTask from './controller/task/createTask'
-import getTaskById from './controller/task/getTaskById'
-import login from './controller/user/login'
-import { connection } from './data/connection'
 
 
 dotenv.config()
+
+export const connection = knex({
+   client: 'mysql',
+   connection: {
+      host: process.env.DB_HOST,
+      user: process.env.DB_USER,
+      password: process.env.DB_PASSWORD,
+      database: process.env.DB_NAME,
+      port: 3306
+   }
+})
+
 
 const app = express()
 app.use(express.json())
@@ -21,15 +27,6 @@ app.get("/", async function(req,res){
 })
 
 
-app.post('/user/signup', createUser)
-app.post("/user/login", login)
-app.get('/user/:id', getUserById)
-app.post('/user/edit', editUser)
-
-app.put('/task', createTask)
-app.get('/task/:id', getTaskById)
-
 app.listen(3003, () => {
    console.log('Servidor rodando na porta 3003')
 })
-
