+++ semana19/aula1/src/services/authenticator.ts
@@ -1,5 +1,5 @@
 import * as jwt from "jsonwebtoken"
-import { USER_ROLES } from "../data/user/insertUser"
+import { USER_ROLES } from "../data/insertUser"
 
 export type AuthenticationData = {
    id: string,
